Create calculation_resolver.rb
